# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AidFlow API",
    "description": "AidFlow — 災害情境驅動的災情通報、案件處理、派工與視覺化平台生成器。公開端點位於 /v1/public；其餘為政府後台端點。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/public/platforms/{slug}/cases": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Public case list",
        "operationId": "list_cases_v1_public_platforms__slug__cases_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "phase",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "pending | active | done | open",
              "title": "Phase"
            },
            "description": "pending | active | done | open"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "town",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Town"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "since_hours",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 720,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since Hours"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_desc",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCaseListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/public/platforms/{slug}/cases/{case_id}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Public case detail + timeline",
        "operationId": "get_case_v1_public_platforms__slug__cases__case_id__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/platforms/{platform_id}/cases": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Case queue",
        "operationId": "list_cases_v1_platforms__platform_id__cases_get",
        "parameters": [
          {
            "name": "platform_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Platform Id"
            }
          },
          {
            "name": "phase",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "pending | active | done | open",
              "title": "Phase"
            },
            "description": "pending | active | done | open"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "comma separated statuses",
              "title": "Status"
            },
            "description": "comma separated statuses"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "town",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Town"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "since_hours",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 720,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since Hours"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "created_desc | created_asc | severity_desc | reports_desc | updated_desc",
              "default": "created_desc",
              "title": "Sort"
            },
            "description": "created_desc | created_asc | severity_desc | reports_desc | updated_desc"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Case detail (internal)",
        "operationId": "get_case_v1_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/transition": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Move the case along the state machine",
        "operationId": "transition_v1_cases__case_id__transition_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/assign": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "縣府確認並派工",
        "operationId": "assign_v1_cases__case_id__assign_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/updates": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Add a progress note (public or internal)",
        "operationId": "add_update_v1_cases__case_id__updates_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/photos": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Agency photo (before / scene / after)",
        "operationId": "upload_case_photo_v1_cases__case_id__photos_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_case_photo_v1_cases__case_id__photos_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reports/{report_id}": {
      "get": {
        "tags": [
          "cases",
          "reports"
        ],
        "summary": "Report detail (internal, includes PII)",
        "operationId": "get_report_v1_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Report V1 Reports  Report Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reports/{report_id}/reject": {
      "post": {
        "tags": [
          "cases",
          "reports"
        ],
        "summary": "Mark a report as not credible",
        "operationId": "reject_report_v1_reports__report_id__reject_post",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          },
          {
            "name": "note",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Note"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportInternal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseDetailResponse": {
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseItem"
          },
          "reports": {
            "items": {
              "$ref": "#/components/schemas/ReportInternal"
            },
            "type": "array",
            "title": "Reports"
          },
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/AssignmentItem"
            },
            "type": "array",
            "title": "Assignments"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/CaseEventItem"
            },
            "type": "array",
            "title": "Events"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoItem"
            },
            "type": "array",
            "title": "Photos"
          },
          "nearby": {
            "items": {
              "$ref": "#/components/schemas/NearbyCase"
            },
            "type": "array",
            "title": "Nearby"
          },
          "reporter_roles": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Reporter Roles"
          }
        },
        "type": "object",
        "required": [
          "case",
          "reports",
          "assignments",
          "events",
          "photos",
          "nearby",
          "reporter_roles"
        ],
        "title": "CaseDetailResponse"
      },
...
      "CaseListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "CaseListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicCaseDetail": {
        "properties": {
          "case": {
            "$ref": "#/components/schemas/PublicCase"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/PublicTimelineItem"
            },
            "type": "array",
            "title": "Timeline"
          },
          "reports": {
            "items": {
              "$ref": "#/components/schemas/ReportPublic"
            },
            "type": "array",
            "title": "Reports"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoItem"
            },
            "type": "array",
            "title": "Photos"
          },
          "progress": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Progress"
          }
        },
        "type": "object",
        "required": [
          "case",
          "timeline",
          "reports",
          "photos",
          "progress"
        ],
        "title": "PublicCaseDetail"
      },
      "PublicCaseListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicCase"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "PublicCaseListResponse"
      },
...
      "TransitionRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "目標狀態（須為允許的轉換）"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "public": {
            "type": "boolean",
            "title": "Public",
            "description": "是否顯示在公開時間軸",
            "default": true
          },
          "actor_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Name"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TransitionRequest"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "設定 ADMIN_API_KEY 後，/v1/public 與 /v1/health 以外的端點需帶此標頭。"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}